+++ Sources/Core/AdvancedOperation.swift
@@ -24,20 +24,3 @@
 import Foundation
 
 internal let identifier = "org.tinrobots.AdvancedOperation"
-
-// Quality of Service Inference and Promotion
-//
-// https://developer.apple.com/library/archive/documentation/Performance/Conceptual/EnergyGuide-iOS/PrioritizeWorkWithQoS.html#//apple_ref/doc/uid/TP40015243-CH39
-//
-// Simple and Reliable Threading with NSOperation
-//
-// https://developer.apple.com/library/archive/technotes/tn2109/_index.html#//apple_ref/doc/uid/DTS40010274-CH1-SUBSECTION11
-//
-// Thread explosion
-//
-// https://developer.apple.com/videos/play/wwdc2015/718/
-// https://forums.developer.apple.com/thread/51952
-//
-// Additional Tech Notes
-//
-// https://developer.apple.com/videos/play/wwdc2017/706/