+++ Tests/AdvancedOperationTests/Operations/GroupOperationTests.swift
@@ -35,15 +35,9 @@ final class GroupOperationTests: XCTestCase {
     let operation1 = SleepyAsyncOperation()
     let operation2 = SleepyAsyncOperation()
     let operation3 = SleepyAsyncOperation()
-    
-    operation1.installLogger()
-    operation2.installLogger()
-    operation3.installLogger()
-    
     let groupOperation = GroupOperation(operations: [operation1, operation2, operation3])
     groupOperation.qualityOfService = .userInitiated
-    groupOperation.installLogger()
-    
+
     let expectation1 = XCTKVOExpectation(keyPath: #keyPath(Operation.isFinished), object: groupOperation, expectedValue: true)
     groupOperation.start()
     wait(for: [expectation1], timeout: 5)
@@ -55,7 +49,6 @@ final class GroupOperationTests: XCTestCase {
   
   func testAddingOperationAfterExecution() {
     let groupOperation = GroupOperation(operations: [])
-    groupOperation.installLogger()
     let operation = BlockOperation {
       XCTFail("It shouldn't be executed because GroupOperation is already finished at this point")
     }
@@ -74,18 +67,10 @@ final class GroupOperationTests: XCTestCase {
     let operation2 = SleepyAsyncOperation()
     let operation3 = SleepyAsyncOperation()
     let operation4 = SleepyAsyncOperation()
-    
-    operation1.installLogger()
-    operation2.installLogger()
-    operation3.installLogger()
-    operation4.installLogger()
-    
     let groupOperation1 = GroupOperation(operations: [operation1, operation2])
     let groupOperation2 = GroupOperation(operations: [operation3, operation4])
     let groupOperation3 = GroupOperation(operations: [groupOperation1, groupOperation2])
     groupOperation3.qualityOfService = .userInitiated
-    groupOperation3.installLogger()
-    
     let expectation1 = XCTKVOExpectation(keyPath: #keyPath(Operation.isFinished), object: groupOperation3, expectedValue: true)
     groupOperation3.start()
     wait(for: [expectation1], timeout: 5)
@@ -99,15 +84,10 @@ final class GroupOperationTests: XCTestCase {
     let operation1 = SleepyAsyncOperation(interval1: 1, interval2: 0, interval3: 0)
     let operation2 = SleepyAsyncOperation(interval1: 1, interval2: 0, interval3: 0)
     let operation3 = BlockOperation()
-    
-    operation1.installLogger()
-    operation2.installLogger()
-    operation3.installLogger()
-    
+
     let groupOperation = GroupOperation(operations: [operation1, operation2, operation3])
     groupOperation.maxConcurrentOperationCount = 1
-    groupOperation.installLogger()
-    
+
     let expectation1 = XCTKVOExpectation(keyPath: #keyPath(Operation.isFinished), object: groupOperation, expectedValue: true)
     groupOperation.start()
     wait(for: [expectation1], timeout: 15)
@@ -121,13 +101,7 @@ final class GroupOperationTests: XCTestCase {
     let operation1 = RunUntilCancelledAsyncOperation()
     let operation2 = SleepyAsyncOperation()
     let operation3 = SleepyAsyncOperation()
-    
-    operation1.installLogger()
-    operation2.installLogger()
-    operation3.installLogger()
-    
     let groupOperation = GroupOperation(operations: [operation1, operation2, operation3])
-    groupOperation.installLogger()
     let expectation1 = XCTKVOExpectation(keyPath: #keyPath(Operation.isFinished), object: groupOperation, expectedValue: true)
     groupOperation.start()
     groupOperation.cancel()
@@ -137,22 +111,14 @@ final class GroupOperationTests: XCTestCase {
   func testAddingOperationWhileGroupOperationIsBeingCancelled() {
     let operation1 = RunUntilCancelledAsyncOperation()
     let operation2 = SleepyAsyncOperation(interval1: 1, interval2: 0, interval3: 0)
-    let operation3 = InfiniteOperation()
+    let operation3 = InfiniteAsyncOperation()
     let operation4 = CancelledOperation()
-    
-    operation1.installLogger()
-    operation2.installLogger()
-    operation3.installLogger()
-    operation4.installLogger()
-    
     operation1.name = "op1"
     operation2.name = "op2"
     operation3.name = "op3"
     operation4.name = "op4"
     
     let groupOperation = GroupOperation(operations: [operation1, operation2, operation3])
-    groupOperation.installLogger()
-    
     // cancellation is done while at least one operation is running
     operation3.onExecutionStarted = { [unowned groupOperation] in
       groupOperation.cancel()
@@ -173,13 +139,7 @@ final class GroupOperationTests: XCTestCase {
     let operation1 = SleepyAsyncOperation()
     let operation2 = SleepyAsyncOperation()
     let operation3 = SleepyAsyncOperation()
-    
-    operation1.installLogger()
-    operation2.installLogger()
-    operation3.installLogger()
-    
     let groupOperation = GroupOperation(operations: [operation1, operation2, operation3])
-    groupOperation.installLogger()
     let expectation1 = XCTKVOExpectation(keyPath: #keyPath(Operation.isFinished), object: groupOperation, expectedValue: true)
     groupOperation.cancel()
     groupOperation.start()
@@ -191,13 +151,7 @@ final class GroupOperationTests: XCTestCase {
     let operation1 = SleepyAsyncOperation()
     let operation2 = SleepyAsyncOperation()
     let operation3 = SleepyAsyncOperation()
-    
-    operation1.installLogger()
-    operation2.installLogger()
-    operation3.installLogger()
-    
     let groupOperation = GroupOperation(operations: [operation1, operation2, operation3])
-    groupOperation.installLogger()
     let expectation1 = XCTKVOExpectation(keyPath: #keyPath(Operation.isFinished), object: groupOperation, expectedValue: true)
     queue.addOperation(groupOperation)
     wait(for: [expectation1], timeout: 5)
@@ -221,15 +175,7 @@ final class GroupOperationTests: XCTestCase {
     let operation1 = SleepyAsyncOperation()
     let operation2 = SleepyAsyncOperation()
     let operation3 = SleepyAsyncOperation()
-    let operation4 = SleepyAsyncOperation()
-    
-    operation1.installLogger()
-    operation2.installLogger()
-    operation3.installLogger()
-    operation4.installLogger()
-    
     let groupOperation = GroupOperation(operations: [operation1, operation2, operation3])
-    groupOperation.installLogger()
     let expectation1 = XCTKVOExpectation(keyPath: #keyPath(Operation.isFinished), object: groupOperation, expectedValue: true)
     groupOperation.cancel()
     queue.addOperation(groupOperation)