+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### 5.3.0
+
+- The `isReady` property is not participating anymore in the AsyncOperation internal state change.
+- A `NSException` will be thrown if an AsyncOperation gets started without being ready.
+- A `NSException` will be thrown if an AsyncOperation gets started multiple times.
+- Added `ObservableOperation` to easily observe all KVC compliant Operation properties.
+- Removed `LoggableOperation`.
+
 ### 5.2.0
 
 - Improved the underlying implementawtion of  `GroupOperation`.